+++ src/simulators/manager.ts
@@ -2,7 +2,12 @@ import events from "node:events";
 import { type SimulatorOutput, getSimulators } from "../common/cli/scripts";
 import { commonLogger } from "../common/logger";
 import { assertUnreachable } from "../common/types";
-import { type SimulatorDestination, iOSSimulatorDestination, watchOSSimulatorDestination } from "./types";
+import {
+  type SimulatorDestination,
+  iOSSimulatorDestination,
+  visionOSSimulatorDestination,
+  watchOSSimulatorDestination,
+} from "./types";
 import { parseDeviceTypeIdentifier, parseSimulatorRuntime } from "./utils";
 
 type IEventMap = {
@@ -74,8 +79,16 @@ export class SimulatorsManager {
       return null;
     }
     if (runtime.os === "xrOS") {
-      // todo: add xrOS simulator support
-      return null;
+      return new visionOSSimulatorDestination({
+        udid: simulator.udid,
+        isAvailable: simulator.isAvailable,
+        state: simulator.state as "Booted",
+        name: simulator.name,
+        os: runtime.os,
+        osVersion: runtime.version,
+        rawDeviceTypeIdentifier: simulator.deviceTypeIdentifier,
+        rawRuntime: rawRuntime,
+      });
     }
     assertUnreachable(runtime.os);
   }